Add files via upload
This is an insurance project on which as a data scientist put myself in the shoes of an insurance company. 
The project was initially a challenge at AIcrowd, I was to build a pricing model and compete against other insurance companies for profit.

In the project, I was to compete within the insurance market for profit. The market in this challenge is evaluated as a cheapest-wins market. That means every insurance company offers every customer an annual premium price, and the customer will always pick the company that offers the cheapest price to them (maybe using a price comparison website).

In order to create your pricing model, the project is created using real historical car insurance policies for 4 consecutive years. Each policy concerns 1 vehicle, its drivers, and accident history over 4 years. This data has been provided by a large car insurance provider in a European country and is a uniform sample for their entire portfolio.

By the end of the challenge I was to produce a model to price contracts for incoming policies for the 5th year.
The company that makes the most profit in this market, wins the challenge.

Thanks to my team, Nikhil and Christopher https://www.linkedin.com/in/nikhil-adithyan-32a8b91a5/
https://www.linkedin.com/in/wanjohi-christopher-b9bb9b196/